Time: 76 ms (95.40%), Space: 27 MB (84.53%) - LeetHub

+++ 846-hand-of-straights/846-hand-of-straights.cpp
@@ -1,7 +1,7 @@
 class Solution {
 public:
     bool isNStraightHand(vector<int>& hand, int grpsize) {
-        unordered_map<long long,long long> mp;
+        unordered_map<int,int> mp;
         sort(hand.begin(),hand.end());
         if(hand.size()%grpsize)
             return false;
